+++ Assets/AmplifyShaderEditor.7z.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 52d44cfd8115ba4428b14cbe362029a6
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
