+++ Assets/Scripts/Player/Player.cs
@@ -2,124 +2,6 @@ using UnityEngine;
 using Player_FSM;
 using System;
 using System.Collections;
-using Cinemachine;
-using UnityEngine.Serialization;
-
-
-[Serializable]
-public class WalkSettings
-{
-    [Header("行走")] public float walkSpeed; //行走速度
-
-    public float accelerate; //加速度
-    public float groundDrag; //地面阻力
-}
-[Serializable]
-public class JumpSettings
-{
-    [Header("跳跃")] public float height; //跳跃高度
-    
-    [Header("墙跳")] 
-    public float wallJumpSpeed;
-    public bool isWallJump;
-    public float exitWallTime;
-
-}
-[Serializable]
-public class SprintSettings
-{
-    [Header("冲刺")] public float sprintSpeed; //冲刺速度
-    public float sprintDistance; //冲刺距离
-    public float sprintLeaveSpeed;//冲刺结束的速度
-    
-    public float sprintTime; //冲刺时间(无需暴露)
-}
-
-[Serializable]
-public class CrouchSettings
-{
-    [Header("下蹲")] public float crouchSpeed; //蹲行速度
-    public float crouchYScale; //下蹲时y缩放量
-}
-
-[Serializable]
-public class SlidingSettings
-{
-    [Header("滑行")] public float maxSlideTime; //最大滑行时间
-    public float vineLineTime; //速度线时间
-    public float slideYScale; //滑行时y缩放度
-    public float slideAccelerate; //滑行减速的加速度
-    public float startSlideSpeed = 1f; //滑行至少需要的速度
-}
-
-[Serializable]
-public class AirSettings
-{
-    public float airTransformAccelerate;
-}
-[Serializable]
-public class WallRunningSettings
-{
-    [Header("贴墙跑")] 
-    //墙跑速度改为继承
-    public float wallRunSpeed; //墙跑速度
-    public LayerMask whatIsWall; //wall的图层
-    public float wallRunGRate = 0.1f; //滑墙重力倍率
-
-    //public float maxWallTime; //最大墙跑时间
-    public float wallCheckDistance; //墙跑检测距离
-    public float wallRunMinDisTance;
-    public RaycastHit wallLeftHit;
-    public RaycastHit wallRightHit;
-
-    public RaycastHit currentWall;
-
-    public bool rightWall;
-    public bool leftWall;
-}
-
-
-
-[Serializable]
-public class KeySettings
-{
-    [Header("按键")] public KeyCode jumpkey = KeyCode.Space; //跳跃按键
-    public KeyCode sprintKey = KeyCode.LeftShift; //冲刺按键
-    public KeyCode crouchKey = KeyCode.C; //下蹲按键
-    public KeyCode slideKey = KeyCode.LeftControl; //滑行按键
-}
-[Serializable]
-public class OtherSettings
-{
-    //太繁琐,待自定义editor
-    [Header("游戏状态")] 
-    public Vector3 dirInput;
-    public Vector3 moveDir;
-    public Rigidbody m_rigidbody;
-    public Transform orientation;
-    public Vector3 speed = new Vector3(0, 0, 0); //继承速度
-    public float speedMag; //当前动量;
-    public Transform camTrans;
-    public Camera cam;
-    public EStateType last;
-    public EStateType current;
-    public EStateType next;
-
-    [Header("着地检测")] public float playerHeight; //玩家最低高度
-    public LayerMask whatIsGround; //地面图层
-
-    [Header("上坡")] public float maxSlopeAngle; //最大坡度
-
-    [Header("逻辑变量")] 
-    public float sprintChangeRate; //冲刺动量转化速率
-    public float walkToSlideCovoteTime; //walk状态转化为slide的动量继承土狼时间;
-    public float slideToJumpHeightRate = 0.5f; //slide到jump状态,跳跃高度的变化率
-
-    [Space(10)] [Header("武器")] public Transform gunModel;
-    public Transform gunTrans;
-
-    [Header("其他效果")] public VLineSummon vineLine;
-}
 
 [Serializable]
 public class PlayerBlackboard : Blackboard
