+++ Assets/Scripts/Player/PlayerSettings.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2c37d736d9f1c664aa5fad5b00041d02
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
