+++ Assets/Scripts/Player/PlayerSettings/AirSettings.cs
@@ -1,0 +1,7 @@
+using System;
+
+[Serializable]
+public class AirSettings
+{
+    public float airTransformAccelerate;
+}
