+++ Assets/Scripts/Player/PlayerSettings/AirSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8bad55db2ac828c4591c3bc5ceb9794b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
