+++ Assets/Scripts/Player/PlayerSettings/CrouchSettings.cs
@@ -1,0 +1,9 @@
+using System;
+using UnityEngine;
+
+[Serializable]
+public class CrouchSettings
+{
+    [Header("下蹲")] public float crouchSpeed; //蹲行速度
+    public float crouchYScale; //下蹲时y缩放量
+}
