+++ Assets/Scripts/Player/PlayerSettings/CrouchSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5b364da576054b24f9fcf3a8708d0ff5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
