+++ Assets/Scripts/Player/PlayerSettings/JumpSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c07e8a73612e15242b13b243fef27c64
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
