+++ Assets/Scripts/Player/PlayerSettings/KeySettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4098924a72a1dde46a4f57a307915e33
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
