+++ Assets/Scripts/Player/PlayerSettings/OtherSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d2a2039f9c635674484f6e54d82bb1bb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
