+++ Assets/Scripts/Player/PlayerSettings/SlidingSettings.cs
@@ -1,0 +1,12 @@
+using System;
+using UnityEngine;
+
+[Serializable]
+public class SlidingSettings
+{
+    [Header("滑行")] public float maxSlideTime; //最大滑行时间
+    public float vineLineTime; //速度线时间
+    public float slideYScale; //滑行时y缩放度
+    public float slideAccelerate; //滑行减速的加速度
+    public float startSlideSpeed = 1f; //滑行至少需要的速度
+}
