+++ Assets/Scripts/Player/PlayerSettings/SlidingSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8a36ffc64b4c2734e8c509005ae6cd51
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
