+++ Assets/Scripts/Player/PlayerSettings/SprintSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2b167743466ec1e4593f95cb17dd51c1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
