+++ Assets/Scripts/Player/PlayerSettings/WalikSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 062eb1fea9d14544db8bd4cb417b65b0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
