+++ Assets/Scripts/Player/PlayerSettings/WallRunningSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 03c57845a2e76d74eaf24f057d893997
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
